# apps/api/openapi.gen.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Hyprnote API",
    "description": "Development documentation",
    "version": "1.0.0"
  },
...
  "components": {
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    },
...
  },
  "servers": [
    {
      "url": "https://api.hyprnote.com",
      "description": "Production server"
    },
    {
      "url": "http://localhost:4000",
      "description": "Local development server"
    }
  ],
  "paths": {
...
    "/listen": {
      "get": {
        "responses": {
          "101": {
            "description": "-"
          },
          "400": {
            "description": "-"
          },
          "401": {
            "description": "-"
          },
          "502": {
            "description": "-"
          },
          "504": {
            "description": "-"
          }
        },
        "operationId": "getListen",
        "tags": ["private-skip-openapi"],
        "parameters": [],
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/transcribe": {
      "post": {
        "responses": {
          "200": {
            "description": "result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "metadata": {},
                    "results": {
                      "type": "object",
                      "properties": {
                        "channels": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "alternatives": {
                                "type": "array",
                                "items": {
                                  "type": "object",
                                  "properties": {
                                    "transcript": {
                                      "type": "string"
                                    },
                                    "confidence": {
                                      "type": "number"
                                    },
                                    "words": {
                                      "type": "array",
                                      "items": {
                                        "type": "object",
                                        "properties": {
                                          "word": {
                                            "type": "string"
                                          },
                                          "start": {
                                            "type": "number"
                                          },
                                          "end": {
                                            "type": "number"
                                          },
                                          "confidence": {
                                            "type": "number"
                                          },
                                          "speaker": {
                                            "anyOf": [
                                              {
                                                "type": "number"
                                              },
                                              {
                                                "type": "null"
                                              }
                                            ]
                                          },
                                          "punctuated_word": {
                                            "anyOf": [
                                              {
                                                "type": "string"
                                              },
                                              {
                                                "type": "null"
                                              }
                                            ]
                                          }
                                        },
                                        "required": ["word", "start", "end", "confidence"],
                                        "additionalProperties": false
                                      }
                                    }
                                  },
                                  "required": ["transcript", "confidence", "words"],
                                  "additionalProperties": false
                                }
                              }
                            },
                            "required": ["alternatives"],
                            "additionalProperties": false
                          }
                        }
                      },
                      "required": ["channels"],
                      "additionalProperties": false
                    }
                  },
                  "required": ["metadata", "results"],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "-"
          },
          "401": {
            "description": "-"
          },
          "500": {
            "description": "-"
          },
          "502": {
            "description": "-"
          }
        },
        "operationId": "postTranscribe",
        "tags": ["private-skip-openapi"],
        "parameters": [],
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
...
  }
}